Test Fixes - portlet/blogs/blogsentry
