+++ portal-web/test/functional/com/liferay/portalweb/portlet/blogs/blogsentry/addblogsentries/AddBlogsEntriesTests.html
