+++ portal-web/test/functional/com/liferay/portalweb/portlet/blogs/blogsentry/addblogsentries/ViewBlogsEntry1Test.html
